[77] --> series of patches
-fix patch 1
-number one fiber
-merge 5.5.2013
-up from git://github.com/ChaosDevLikan/multi_realm_cell.git
-up from Goliath Data center home server
http://127.0.0.1/ChaosDevLikan/multi_realm_cell.git
-up from git://github.com/TrinityCore/TrinityCore.git
-fix patch 2
-buffering on
-many fixes
-lua fixes
-new technology codes
-fix patch 3
-add newest:
< ace/bzip2/g3dlite/gsoap/jemalloc/libmpq/recastnavigation > in dep_tc
-add old sistem Errors in debugging
-fix compiling
-new spells
-new auras
-new opcodes
-newest movement
-some taktiks
-prepare for the new patch 4

+++ src/src_cata/server/LuaEngine/LuaHooks.cpp
@@ -850,7 +850,7 @@ public:
             sEluna->ExecuteCall(3, 0);
         }
     }
-    void OnMoneyChanged(Player* player, int32& amount)
+    void OnMoneyChanged(Player* player, int64& amount)
     {
         for (std::vector<int>::iterator itr = sEluna->ServerEventBindings[PLAYER_EVENT_ON_MONEY_CHANGE].begin();
             itr != sEluna->ServerEventBindings[PLAYER_EVENT_ON_MONEY_CHANGE].end(); ++itr)
@@ -1169,7 +1169,7 @@ public:
             sEluna->ExecuteCall(2, 0);
         }
     }
-    void OnMemberWitdrawMoney(Guild* guild, Player* player, uint32 &amount, bool isRepair)
+    void OnMemberWitdrawMoney(Guild* guild, Player* player, uint64 &amount, bool isRepair)
     {
         for (std::vector<int>::iterator itr = sEluna->ServerEventBindings[GUILD_EVENT_ON_MONEY_WITHDRAW].begin();
             itr != sEluna->ServerEventBindings[GUILD_EVENT_ON_MONEY_WITHDRAW].end(); ++itr)
@@ -1183,7 +1183,7 @@ public:
             sEluna->ExecuteCall(5, 0);
         }
     }
-    void OnMemberDepositMoney(Guild* guild, Player* player, uint32 &amount)
+    void OnMemberDepositMoney(Guild* guild, Player* player, uint64 &amount)
     {
         for (std::vector<int>::iterator itr = sEluna->ServerEventBindings[GUILD_EVENT_ON_MONEY_DEPOSIT].begin();
             itr != sEluna->ServerEventBindings[GUILD_EVENT_ON_MONEY_DEPOSIT].end(); ++itr)
